modulized grasping client
